requirements.txt
asgiref==3.2.3
Click==7.0
Django==3.0.3
django-ckeditor==5.9.0
django-filter==2.2.0
django-js-asset==1.2.2
django-tempus-dominus==5.1.2.13
Pillow==7.0.0
pip-tools==4.4.1
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
